// namma-bites/src/app/core/models/enums.ts
export enum CuisineType {
  Andhra = 'Andhra',
  Biryani = 'Biryani',
  NorthIndian = 'North Indian',
  SouthIndian = 'South Indian',
  Chinese = 'Chinese',
  Continental = 'Continental',
  FastFood = 'Fast Food',
  Desserts = 'Desserts',
  Beverages = 'Beverages',
  Cafe = 'Cafe',
  Grill = 'Grill',
  Bakery = 'Bakery',
  StreetFood = 'Street Food'
}

export enum FoodType {
  Veg = 'VEG',
  NonVeg = 'NON_VEG',
  Egg = 'EGG'
}

export enum OrderStatus {
  Placed = 'PLACED',
  Confirmed = 'CONFIRMED',
  Preparing = 'PREPARING',
  OutForDelivery = 'OUT_FOR_DELIVERY',
  Delivered = 'DELIVERED',
  Cancelled = 'CANCELLED'
}

export enum PaymentMethod {
  COD = 'COD',
  UPI = 'UPI',
  Card = 'CARD'
}